third_party/WebKit/LayoutTests/inspector/console/console-uncaught-promise-expected.txt

CONSOLE ERROR: Unhandled promise rejection
CONSOLE ERROR: Unhandled promise rejection
CONSOLE ERROR: Unhandled promise rejection
CONSOLE ERROR: Unhandled promise rejection
Tests that uncaught promise rejections are logged into console.

Unhandled promise rejection Promise
    __proto__: Promise
    [[PromiseStatus]]: "rejected"
    [[PromiseValue]]: Error: onload.err1
Unhandled promise rejection Promise
    __proto__: Promise
    [[PromiseStatus]]: "rejected"
    [[PromiseValue]]: Error: onload.err2
Unhandled promise rejection Promise {[[PromiseStatus]]: "rejected", [[PromiseValue]]: Error: inspector.err1}
    __proto__: Promise
    [[PromiseStatus]]: "rejected"
    [[PromiseValue]]: Error: inspector.err1
Unhandled promise rejection Promise {[[PromiseStatus]]: "rejected", [[PromiseValue]]: Error: inspector.err2}
    __proto__: Promise
    [[PromiseStatus]]: "rejected"
    [[PromiseValue]]: Error: inspector.err2
